Read, visualize, filter, and clean data
Adds a number of functions to prepare data for training.

loadData() returns a Promise that will resolve with an array of all data
from the 'colors' database.

filterData(...) returns a dataset free of junk entries, and filters the
data to just one label.

analyzeData(...) simply prints two lines. First, the number of non-junk
entries for that label. Second, the average hue of aforementioned
entries.

showSample(...) clears the canvas and instead displays non-junk entries
of the given label. One color per pixel.

Last, but most important:
cleanData(...) returns an array of data which has been cleaned of junk
entries, limited to one label, and every color falls in the given hue
range.

As a side effect of this commit (and my prettier settings), all strings
are now wrapped by ""