Layout and color selection changes.
Don't display the RGB values to the user to discourage "cheating".

The color is presented to the user twice side by side, once on a black
background and once on a white background. The background of the page
is a gray color midway between black and white. This seems to improve
color distinction.

The buttons have been reordered to more closely align with the color
spectrum.

Choosing from random RGB values biases the colors towards green-ish.
Instead, we want to uniformly sample from all hues.
We bias towards full saturation or else too many colors look either
gray,  brown, or pink.
We bias away from full black and full white primarily because they
aren't  options presented to the user.